don't let the simulation run for one second.
When the environment AntBulletEnv resets, the ant starts in the air. I thought it was idea to let it fall and start the control once it touches
the ground plane. However it seems that the body of the ant touches the ground and then the episode is done.
Analyzing this a bit, this is what happens but during the training I observed episodes with 20 steps and many more at the beggining, even the maximum of 1K.
Apart from that it doesn't seem good to use as starting state the one when the ant is in the air although we start contolling it once it is in the floor.